style: Reduce clippy warnings from 94 to 31
Fixes applied:
- non-canonical partial_cmp: delegate to cmp() instead of reimplementing
- clone to slice_from_ref: use std::slice::from_ref
- Add #[allow(dead_code)] for public APIs not yet used internally:
  - bridge/service/db.rs (Database impl, OptionalExt trait)
  - bridge/service/watchers (BthWatcher, EthereumWatcher structs)
  - crypto/ring-signature (batch verify functions)
  - common/node_id.rs (NodeIDError enum)
  - consensus/scp (logger fields in LoggingScpNode and Slot)
- &Box<T> to &T in exchange-scanner
- Remove unused variables (self_ids, p)
- Remove unnecessary casts (u64 as u64, u32 as u32)
- Use abs_diff() instead of manual absolute difference
- Use is_multiple_of() instead of modulo comparison
- Use !is_empty() instead of len() > 0
- Fix doc list item indentation
- Fix empty line after doc comment

Remaining 31 warnings are mostly:
- Missing documentation in gossip crate (internal structs)
- Loop variable indexing suggestions
- Method naming conventions (API design choices)

+++ botho/src/block.rs
@@ -868,7 +868,7 @@ pub mod difficulty {
             let net_per_tx = self.total_emitted.saturating_sub(self.total_burned)
                 / self.total_tx;
             let annual = net_per_tx * 10_000_000;
-            (annual * 10_000 / supply) as u64
+            annual * 10_000 / supply
         }
     }
